+++ src/Service/Validator.h
@@ -1,0 +1,10 @@
+#pragma once
+#include "../Common/Protocol.h"
+#include <mutex>
+
+namespace dowe::service {
+class Validator {
+public: ipc::Response Validate(const ipc::Request& request) noexcept;
+private: std::mutex mutex_;
+};
+} // namespace dowe::service